ISSUES-#362: Move the 'useAddressForm' hook to under folder of 'AddressForm'
* Change the path of import in 'addressFormContainer.js' to reflect the new path of the hook

* Move the corresponding test file as well
